refactor(data): align DTOs, mappers, datasources, and repositories with backend audit
- Add AddressUpdateRequest, AppConfigUpsertRequest, FoodBulkStatusRequest,
  VillageAddRequest DTOs
- Update DomainMapper, EntityMapper, RequestMapper for new fields
- Update all response DTOs (OrderResponse, CafeAddressResponse, etc.)
- Align ApiEndPoint with backend (address, village, media, bulk-status endpoints)
- Update all RemoteDataSource implementations
- Update all RepositoryImpl classes with new method signatures
- Update DataUiMock with additional mock data

+++ data/src/main/java/com/mtv/app/shopme/data/remote/response/AppConfigResponse.kt
@@ -12,8 +12,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class AppConfigResponse(
-    val minVersion: Int = 0,
-    val latestVersion: Int = 0,
+    val minVersion: Long = 0,
+    val latestVersion: Long = 0,
     val forceUpdate: Boolean = false,
     val maintenanceMode: Boolean = false,
     val maintenanceMessage: String? = null
